+++ atst/app.py
@@ -14,6 +14,7 @@ from atst.filters import register_filters
 from atst.routes import bp
 from atst.routes.portfolios import portfolios_bp as portfolio_routes
 from atst.routes.task_orders import task_orders_bp
+from atst.routes.applications import applications_bp
 from atst.routes.dev import bp as dev_routes
 from atst.routes.users import bp as user_routes
 from atst.routes.errors import make_error_pages
@@ -71,6 +72,7 @@ def make_app(config):
     app.register_blueprint(bp)
     app.register_blueprint(portfolio_routes)
     app.register_blueprint(task_orders_bp)
+    app.register_blueprint(applications_bp)
     app.register_blueprint(user_routes)
 
     if ENV != "prod":
