+++ atst/routes/applications/settings.py
@@ -1,0 +1,76 @@
+from flask import redirect, render_template, request as http_request, url_for
+
+from . import applications_bp
+from atst.domain.environment_roles import EnvironmentRoles
+from atst.domain.applications import Applications
+from atst.forms.application import ApplicationForm
+from atst.domain.authz.decorator import user_can_access_decorator as user_can
+from atst.models.permissions import Permissions
+from atst.utils.flash import formatted_flash as flash
+
+
+def get_environments_obj_for_app(application):
+    environments_obj = {}
+
+    for env in application.environments:
+        environments_obj[env.name] = []
+        for user in env.users:
+            env_role = EnvironmentRoles.get(user.id, env.id)
+            environments_obj[env.name].append(
+                {"name": user.full_name, "role": env_role.displayname}
+            )
+
+    return environments_obj
+
+
+@applications_bp.route("/applications/<application_id>/settings")
+@user_can(Permissions.VIEW_APPLICATION, message="view application edit form")
+def settings(application_id):
+    application = Applications.get(application_id)
+    form = ApplicationForm(name=application.name, description=application.description)
+
+    return render_template(
+        "portfolios/applications/edit.html",
+        application=application,
+        form=form,
+        environments_obj=get_environments_obj_for_app(application=application),
+    )
+
+
+@applications_bp.route("/applications/<application_id>/edit", methods=["POST"])
+@user_can(Permissions.EDIT_APPLICATION, message="update application")
+def update(application_id):
+    application = Applications.get(application_id)
+    form = ApplicationForm(http_request.form)
+    if form.validate():
+        application_data = form.data
+        Applications.update(application, application_data)
+
+        return redirect(
+            url_for(
+                "applications.portfolio_applications",
+                portfolio_id=application.portfolio_id,
+            )
+        )
+    else:
+        return render_template(
+            "portfolios/applications/edit.html",
+            application=application,
+            form=form,
+            environments_obj=get_environments_obj_for_app(application=application),
+        )
+
+
+@applications_bp.route("/applications/<application_id>/delete", methods=["POST"])
+@user_can(Permissions.DELETE_APPLICATION, message="delete application")
+def delete(application_id):
+    application = Applications.get(application_id)
+    Applications.delete(application)
+
+    flash("application_deleted", application_name=application.name)
+
+    return redirect(
+        url_for(
+            "applications.portfolio_applications", portfolio_id=application.portfolio_id
+        )
+    )
