+++ atst/routes/portfolios/__init__.py
@@ -4,7 +4,6 @@ from operator import attrgetter
 portfolios_bp = Blueprint("portfolios", __name__)
 
 from . import index
-from . import applications
 from . import members
 from . import invitations
 from . import task_orders
