+++ templates/portfolios/applications/index.html
@@ -13,7 +13,7 @@
     <div class='portfolio-applications__header--title col col--grow'>Applications</div>
     <div class='portfolio-applications__header--actions col'>
       {% if can_create_applications %}
-        <a class='icon-link' href='{{ url_for('portfolios.new_application', portfolio_id=portfolio.id) }}'>
+        <a class='icon-link' href='{{ url_for('applications.new', portfolio_id=portfolio.id) }}'>
           {{ 'portfolios.applications.add_application_text' | translate }}
           {{ Icon("plus", classes="sidenav__link-icon icon--circle") }}
         </a>
@@ -26,7 +26,7 @@
     {{ EmptyState(
       'This portfolio doesn’t have any applications yet.',
       action_label='Add a new application' if can_create_applications else None,
-      action_href=url_for('portfolios.new_application', portfolio_id=portfolio.id) if can_create_applications else None,
+      action_href=url_for('applications.new', portfolio_id=portfolio.id) if can_create_applications else None,
       icon='cloud',
       sub_message=None if can_create_applications else 'Please contact your JEDI Cloud portfolio administrator to set up a new application.'
     ) }}
@@ -43,14 +43,14 @@
                 <span class='accordion__description'>{{ application.description }}</span>
                 <div class='accordion__actions'>
                   {% if user_can(permissions.VIEW_APPLICATION) %}
-                    <a class='icon-link' href='{{ url_for("portfolios.edit_application", portfolio_id=portfolio.id, application_id=application.id) }}'>
+                    <a class='icon-link' href='{{ url_for("applications.settings", application_id=application.id) }}'>
                       <span>{{ "portfolios.applications.app_settings_text" | translate }}</span>
                     </a>
                     <div class='separator'></div>
                   {% endif %}
                   {% if user_can(permissions.VIEW_PORTFOLIO_USERS) %}
                   <a
-                    href="{{ url_for('portfolios.application_team', portfolio_id=portfolio.id, application_id=application.id) }}"
+                    href="{{ url_for('applications.team', application_id=application.id) }}"
                     class='icon-link'>
                       <span>{{ "portfolios.applications.team_text" | translate }}</span>
                       <span class='counter'>{{ application.num_users }}</span>
@@ -76,7 +76,7 @@
                     <span>{{ environment.name }}</span>
                   </div>
 
-                  <a href='{{ url_for("portfolios.access_environment", portfolio_id=portfolio.id, environment_id=environment.id)}}' target='_blank' rel='noopener noreferrer' class='application-list-item__environment__csp_link icon-link'>
+                  <a href='{{ url_for("applications.access_environment", environment_id=environment.id)}}' target='_blank' rel='noopener noreferrer' class='application-list-item__environment__csp_link icon-link'>
                     <span>{{ "portfolios.applications.csp_console_text" | translate }}</span>
                   </a>
